// src/app/modules/configuration/finance/models/fee-acccounts-search-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * FeeAccountDataView
 */
export class FeeAccountsSearchView { 

    sortBy?: string;
    sortOrder?: number;
    pageNumber: number;
    pageSize: number;
    accountName?: Array<string>;
    failedRecords?: Array<string>;
    displayName?: Array<string>;
    code?: Array<string>;
    description?: Array<string>;
}